Extract WAV normalization actions
